network: splitting firewall policy into two files
